add back init route field mods
seems like at some point the code to mutate some of the default route fields
such as 'memory', 'format', 'timeout', 'idle_timeout' got removed, I'd guess
in the name of simplifying the func file output which I agree with, but would
be nice to modify these fields too (or perhaps a mistake?). we could omit them
if the defaults are captured, but that means we have to sync with the fn
server so not a huge fan of that (already an issue, really). anyway, the
defaults are back to being spat out, unless modified via flags.

found the ascii text pretty obnoxious in init, removed it. just created a
bunch of empty space in my shell that i had to scroll over, provides no value.
don't even like having any output when the command succeeds but that's too
neckbeard for this project probably...

now for tutorials we can do:

```
fn init --format http
```

instead of saying "open up your func.yaml in your favorite editor and add a
line...". also, nobody has to do that anymore, which is nice, too. only
works on select fields. i didn't want to do the whole update funcy dance
today, would be nice to embed a models.Route in the funcfile struct but it
doesn't have `yaml` tags and generates ugly fields :/ so, this works for now.