http fn run only print body (#209)
this refactors the run code into separate blocks for handling each format.
it's not great, but it's a start, further massaging is rarely punished.

json was doing this, and seems to be the expectation. now json/http/default
all have the same output for my function in `fn run`.

closes #208

`--runs > 1` didn't work on master, and continues not to work here. I've gotta
switch gears, this at least fixes the http kerfuffle. that whole feature looks
like it should be re-thought or dropped, this code is/was pretty nasty altogether.

master:

```
✗: cli run
Building image rdallman/yoyodawg:0.0.71 .
HTTP/1.1 200 OK
Content-Length: 26

{"message":"Hello World"}
```

now:

```
✗: cli run
Building image rdallman/yoyodawg:0.0.71 .
{"message":"Hello World"}
```