Making config importable
Having the config names defined in the main package meant they
couldn't be reused in other packages, namely client/api.

This creates the config package to contain the config key definitions
and makes them exportable.

Also moved the loadConfiguration func into config from main, and it
feels like a better place for it to live.