Create command checks the file hasn't already been create and creates it either with default values or with values passed in through the --provider and --apiurl flags.
Set checks there is a context file and sets the current context to the context name and
updates the config.yaml file.

List gets all available context files, unmarshals them and appends them to a table.

add protection for the default context so it can not be
deleted and for the current context

include a nicer error message when both FN_API_URL and api_url in a
context file are not set. Also allow for FN_API_URL to override
api_url defined in a context file.

validate the name of the context when creating a context file, using
regex to allow for ASCII, hyphens and dashes.

Initialisation of .fn will create a config.yaml and context/default.yaml but will not set the current
context to default. Instead it will be set to "". A user will have to explicitly use the command:
fn context s default to set the current-context. This will mean there are no changes for current
users and they can essentially opt in to use contexts.

Added a message to notify user when FN_API_URL, current-context and api-url in context file is not set.

removed commented code and print lines

print out api_url warning for all commamnds that use it

add warning for env vars overriding values in context file