Route annotations can be specified in func.yaml. (#290)
* Route annotations can be specified in func.yaml.

Annotations for a route can be specified using `route_annotations` in
the func.yaml. For example:

```
route_annotations:
    a: x
```

Here, `x` can be any valid JSON or YAML value. This allows more
complex annotation values to be specified, such as arrays:

```
route_annotations:
    a:
    - x
    - y
```

Or maps:

```
route_annotations:
    a:
        b: 1
        c: 2
```

* Rename route_annotation field to `annotations`.

This was intended to allow app annotations to be added later, but they
can be added to `app.yaml`.