Functions and Trigger commands (#323)
* initial apps changes for triggers

* basic apps and fns

* inital trigger command work

* wip

* add update command for triggers

* Use constructor function for DeleteFnParams.

Using the struct directly means that the timeout doesn't get set,
which results in a "context deadline exceeded" error.

* Update to use Error within payload.

* Errors are now unwrapped in Swagger and client.

* remove endpoint from fn list

* All apps commands work as expected

* Initial functions

* remove the use of empty object for apps/fn/triggers

* Working crud trigger commands

* update fn_go version in deps

* Update apps error

* Check if registry is passed through on an image on create fn/routes, if not
if registry is set append it to the image and return.

* Change apps back to using v1 client for now to avoid back compat
issues.

Change functions and triggers commands to be singular or plural
on respective commands.

Change the comments for public func commands on functions.

* get the trigger endpoint from the annotations

* initial apps changes for triggers

basic apps and fns

inital trigger command work

wip

add update command for triggers

Use constructor function for DeleteFnParams.

Using the struct directly means that the timeout doesn't get set,
which results in a "context deadline exceeded" error.

Update to use Error within payload.

Errors are now unwrapped in Swagger and client.

remove endpoint from fn list

All apps commands work as expected

Initial functions

remove the use of empty object for apps/fn/triggers

Working crud trigger commands

update fn_go version in deps

Update apps error

Check if registry is passed through on an image on create fn/routes, if not
if registry is set append it to the image and return.

Change apps back to using v1 client for now to avoid back compat
issues.

Change functions and triggers commands to be singular or plural
on respective commands.

Change the comments for public func commands on functions.

get the trigger endpoint from the annotations

* Capitalize usage text on functions and triggers commands

* Capitalized table headings for all list commands.

Added a table format for listing app config

Added a message when no routes/triggers/functions are found when
listing. Similiar to that of apps.

* Capitalized table headings for all list commands.

Added a table format for listing app config

Added a message when no routes/triggers/functions are found when
listing. Similiar to that of apps.

* fix tests

* add tests for functions and triggers crud operations

* Add the creation of app/func for v2 tests.

Change the inspect command on tests to search stdout.

Miss spelling of annotations.

* bump fn_go version