Adding invoke command which uses fn invoke to execute fns
This does not add support for 'fn test', with this change that will
continue to only work with routes. If we merge this we'll have to loop
back and change that over to work with fns.

The url to invoke the fn at comes back as an annotation on the fn metadata.