Adding invoke command which uses fn invoke to execute fns (#379)
This does not add support for 'fn test', with this change that will
continue to only work with routes. If we merge this we'll have to loop
back and change that over to work with fns.

The url to invoke the fn at comes back as an annotation on the fn metadata.

This invocation does not make use of a swagger generated client for the invoke endpoint, but it does use the provider.wrapTransport function to make it work with different providers. If this becomes troublesome we'll have to revisit the invoke swagger.