init-images (#369)
* Initial implementation of init-images

This command:

  fn init --image=fnproj/java-init:latest

Will `docker run` that image, piping the output to `tar -x`. The expected
result is that a template of the function will be created, along with a
func.yaml which is missing `Name` and `Version`.  These are then added by
the cli, so that the function is ready to `fn run`.

Future work is to remove the hard-coded templates in the cli, and make

  fn init --runtime=java

equivalent to the line above. Once this is done for all *supported* runtimes,
the others (php, lambda-node, rust, etc) can be removed and LanguageHelper
is then unnecessary.

Co-Author: vittoria.placidi@oracle.com

* Pass function name to init-image as FN_FUNCTION_NAME in env

  Changes arg name to --init-image based on feedback
  Consistently using init-image and initImage in code and comments

* Change the expected name of func.yaml fragment generated by the init image

  It's confusing to have the init-image generate a file called func.yaml
only to have that read by the CLI and some different content written into
that file.

  This changes the expectation so that the init-image produces a func.init.yaml
and the CLI reads that, merges some more data into it and writes it out as
func.yaml

* Use archive/tar instead of external `tar -x`

  Not reliable that `tar -x` will behave the same everywhere so
we write our own implementation here.

* Adds `init-image` to the code-paths for v2 func files

  Now you can specify `--init-image` with `--trigger` and
it still works.

* Merge func.yaml properties from init-image and defaults, with CLI-overrides

  Almost all the build-related properties should come from the
  init-image.

  Runtime properties should come from defaults (notable exception:
  format)

  Some func.yaml properties are overrideable by CLI args which, if
  used, will probably yield a broken function but that's out of
  scope for this change.

* Adds an integration test for init-images

  As there are no public init images yet, we create one during this
test (by running `fn init` and creating an init-image out of the
result).