remove welcome message
this supposedly runs on new installs, but I:

* don't think it's very useful (not having a changelog withstanding)
* it prints to stdout, which messes up output, which can mess up scripts, etc
* doesn't seem to run as was intended, and doesn't seem like a very good place to be
in case e.g. somebody is piping invoke output to a script... and I can't think
of any place it belongs (we tried to do some cutesy ascii art in `fn init`
once, that was also annoying). open to ideas for better places to put this, i
just can't think of any