fixes default url to not have /v1
it seems we append this now, and including /v2 means we get /v2/v2, and
even xzhibit doesn't like /v2 that much. this gets rid of all the references
to /v1 that were laying around for the api url. guess we forgot to test this,
oops.

also updated the deps with just `dep ensure` -- this means they were stale, I
didn't type update, so this should fix the mismatch there...