Allow fn config to be updated
Currently, the code only allows setting the configuration on a fn, if the configuration is already empty. This ends up preventing the config of an already existing fn being updated, as the fn is pulled down from the server with config present.

Here we remove the check so that the fn's configuration is set to any that has been provided as args. The server side will merge these new values when they are sent up, removing any keys that are set to an empty value.