moving app check up
now that we require apps to exist when deploying them, we should check whether
the app exists before running around and bumping and creating images and all
that, this makes that change. for deploy all this reduces the number of api
calls substantially, we were checking the app each function. obviously this
whole thing is not really atomic, so it can fail pretty spectacularly, but we
aren't really worried about that (and arguably this code is much safer to
delete-create race deploying to 'the wrong app' since it uses id all the way
through for deploy all)

cleaned up error handling to only create functions, triggers, and apps if they
don't exist, not just if any error happens. saves some api calls when service
is down or something, but mostly it doesn't mask any errors from doing the
original get when it's not a not found error (things like permissions this
could be useful).

reduced api calls by one when a function doesn't exist (we don't need to go
get it after creating, the api returns it with the id filled in). /sarcasticwoo