invoke json output, further cleaning
we can get headers out of cli invoke now, huzzah. this does a good amount of
tidying up the client.Invoke function, which was pretty raw, and had some
stale stuff on it even. mainly plumbs output out so that the cli can switch on
its flags for output, instead of passing the output and flags down into the
client invoke function. returning *http.Response always sucks, but at least in
CLI program it will terminate quickly and the conns will be cleaned up, we
handle closing the body fine in the program for now at least though.

removed the block to read the request body into a buffer and limit its size,
that seems like the wrong thing to do overall, it means the cli limits the
input so that the server doesn't error about max size, but the input won't be
complete so a function itself could error (imagine, a 1GB JSON that gets
chopped off at 6MB or whatever) - we should throw this at the server until the
server yells that it's too big, and the server should be limiting this
(there's options that are supposed to be turned on).

I had an idea to just output raw HTTP too, this is easy to add, bringing up
for discussion here first though.

looks like:

```
✗: cli invoke --output=json testapp hello
{
    "body": "{\"message\":\"Hello World\"}\n",
    "headers": {
        "Content-Length": [
            "26"
        ],
        "Content-Type": [
            "text/plain; charset=utf-8"
        ],
        "Date": [
            "Thu, 28 Feb 2019 06:33:33 GMT"
        ],
        "Fn-Call-Id": [
            "01D4SCXRK5180043RZJ0000005"
        ],
        "Fn-Http-Status": [
            "200"
        ]
    },
    "status_code": 200
}
```