no longer support logs and calls
this admittedly is not a great getting started xp, however, notably we've been
pointing users to syslog most of the time anyway in order to debug errors.
there is still the option to start fn in debug mode to get the logs, and we
can discuss that wherever, the last time it was brought up to discuss the only
voices that showed up were Jiminy Cricket.