Alias 'fn delete config' to 'fn unset config' (#454) (#563)
* Alias 'fn delete config' to 'fn unset config' (#454)

Users expect to be able to delete configs using 'fn delete'. Therefore
alias the command so they can use either way.

Fixes #454

* Improve help documentation for config commands (#563)

The help documentation for "fn delete config" was "Manage configurations
for apps and functions". This doesn't explain what it does. Therefore
use the command name in the help documentation to improve the explaination.

Fixes #563