add --version flag to fn start (#575)
e.g. `fn start --version 0.3.690` to start a specific version of
`fnproject/fnserver` docker image

tested ^ and latest. I did not do a lot of validation, but we don't appear to
do much of the other args either (that's my excuse and i'm sticking to it!),
it will fail pretty obviously with 'image not found' if someone specifies the
wrong version and we can't too easily check that without registry calls (we
don't need that), though I concede we could check for semver, latest doesn't
fit that either and is most convenient to keep current behavior. alas.