Make Python images smaller
This patch changes the file mode of the `/python` tree as part of the
build image, which avoids duplicating the pip installed libraries in
the final layer. This makes docker push and pull much faster.

The "Hello world" size for Python 3.8.5 with requirements.txt
containing only "fdk" is now 136MB, compared to 161MB without this
patch. Adding "oci" to requirements.txt brings the image size to
202MB, compared to 292MB without this patch.