Add Python 3.8.5 and make Python images smaller (#613)
* Make boilerplate meet PEP 8 standard

The Python style guide (PEP 8) requires a single space around "=".

* Remove unused func "dockerBuildError" (dead code)

* Add Python 3.8.5 to the list of supported languages

Make it default for "python", because it leads to smaller images.

* Make Python images smaller

This patch changes the file mode of the `/python` tree as part of the
build image, which avoids duplicating the pip installed libraries in
the final layer. This makes docker push and pull much faster.

The "Hello world" size for Python 3.8.5 with requirements.txt
containing only "fdk" is now 136MB, compared to 161MB without this
patch. Adding "oci" to requirements.txt brings the image size to
202MB, compared to 292MB without this patch.